\name{SeqLibR-package}
\alias{SeqLibR-package}
\alias{SeqLibR}
\docType{package}
\title{
  Nucleotide Sequence library for R
}
\description{
 Read and deal with nucleic acid sequences. 
}
\details{
\tabular{ll}{
Package: \tab SeqLibR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-01-18\cr
License: \tab GPL\cr
}
}
\author{
Wim de Leeuw <w.c.deleeuw@uva.nl>
}
\keyword{ package }
